feature: добавляет базу для трат

+++ app/src/main/java/com/example/housekeeper/domain/spend/repository/SpendRepository.kt
@@ -3,6 +3,6 @@ package com.example.housekeeper.domain.spend.repository
 import com.example.housekeeper.domain.spend.Spend
 
 interface SpendRepository {
-    suspend fun isSpendExists(id: Int): Boolean
+    suspend fun isSpendExists(id: Long): Boolean
     suspend fun save(spend: Spend): Result<Spend>
 }
